move part of configureAxisTitleMenu() to axisTitle
move extract of draw-map.js : configureAxisTitleMenu() to axisTitle.js and remove configureAxisSubTitleMenu().
noted in comment :
 configureAxisTitleMenu() and configureAxisSubTitleMenu() implemented
 the axis title menu via (Bootstrap) .popover(); this was replaced by
 components/draw/axis-menu.{js,hbs} in c4a39538..cb6c75b5 and the
 popover configuration was dropped after d80d30df.
