+++ frontend/app/components/draw-map.js
@@ -652,8 +652,6 @@ export default Component.extend(Evented, {
                     selectedFeatures_clear : () => this.get('selectedService').selectedFeatures_clear(),
                     deleteAxisfromAxisIDs,
                     removeAxisMaybeStack,
-                    configureAxisTitleMenu,
-                    configureAxisSubTitleMenu,
                    };
       Ember_set(oa, 'axisApi', axisApi);
     }
@@ -4096,186 +4094,9 @@ getBrushExtents(),
 
     }
 
+    //- moved extract of configureAxisTitleMenu() to axisTitle.js and removed configureAxisSubTitleMenu() 
 
-    /** Setup hover menus over axis titles.
-     * So far used just for Delete
-     * @see based on similar configurejQueryTooltip()
-     */
-    function  configureAxisTitleMenu(block) {
-      let options = me.get('urlOptions'),
-      /** the __data__ of the element triggering the menu was axisName, but is
-       * now block; the axis and stack lookups below could now go more directly
-       * via block. */
-      axisName = block.axisName,
-      /** PerpendicularAxis */
-      dotPlot = options && options.dotPlot,
-      /** The first stage of split axes is enabled by options.splitAxes1,
-       * the remainder by options.splitAxes.
-       * In development, splitAxes1 is enabled by default; in production it is disabled by default. 
-       */
-      splitAxes1 = options && options.splitAxes1 || (config.environment !== 'production');
-      if (trace_gui)
-      console.log("configureAxisTitleMenu", axisName, this, this.outerHTML);
-        let node_ = this;
-        let showMenuFn = me.functionHandle('showMenu', showMenu);
-      node_.onclick = showMenuFn;
-      /** Even though showMenuFn is constant, jQuery.on does : handlers.push(handleObj)
-       * each call, perhaps it avoids duplicate registrations only when selector
-       * is passed.
-       * So node_.onclick is used instead of :
-        $(node_)
-        .on('click', showMenuFn);
-        */
-      /** @param e DOM event */
-      function showMenu(e) {
-        let block = this.__data__;
-        if (block.axis.blocks[0] !== block) {
-          dLog('showMenu', 'data block', block, block.axis.blocks);
-          block = block.axis.blocks[0];
-        }
-        /** defined when called via jQuery.on(click) */
-        let jQueryEventInfo = e.originalEvent && [e.originalEvent.path, e.originalEvent.srcElement, e.handleObj.type];
-        dLog('showMenu', this, axisName, this.__data__, this.parentElement, this.parentElement.parentElement,
-             e, jQueryEventInfo);
-        me.sendAction('selectBlock', block.block);
 
-        /** If the axis-menu is already displayed on a different axis,
-         * reposition it to align with the axis of the clicked block title.
-         */
-        if (me.get('menuAxis') && (me.get('menuAxis') !== block)) {
-          me.set('menuAxis', undefined);
-          later(() => me.set('menuAxis', block));
-        } else {
-          me.set('menuAxis', block);
-        }
-        return false; /* for preventDefault(), stopPropagation() */
-      }
-      if (false) {undefined
-        .popover({
-            trigger : "manual", // hover", // "click focus",
-          sticky: true,
-          delay: {show: 200, hide: 1500},
-          container: 'div#holder',
-          placement : "auto bottom",
-          // title : axisName,
-          html: true,
-          /*
-           *	╳	9587	2573	 	BOX DRAWINGS LIGHT DIAGONAL CROSS
-           *	⇅	8645	21C5	 	UPWARDS ARROW LEFTWARDS OF DOWNWARDS ARROW
-           *	↷	8631	21B7	 	CLOCKWISE TOP SEMICIRCLE ARROW
-           *	⇲	8690	21F2	 	SOUTH EAST ARROW TO CORNER
-           */
-          content : ""
-            + iconButton("DeleteMap", "Delete_" + axisName, "&#x2573;" /*glyphicon-sound-7-1*/, "glyphicon-remove-sign", "#")
-            + iconButton("FlipAxis", "Flip_" + axisName, "&#x21C5;" /*glyphicon-bell*/, "glyphicon-retweet", "#")
-            + 
-            (dotPlot ?
-             iconButton("PerpendicularAxis", "Perpendicular_" + axisName, "&#x21B7;" /*glyphicon-bell*/, "glyphicon-retweet", "#") : "")
-            +
-            (splitAxes1 ?
-             iconButton("ExtendMap", "Extend_" + axisName, "&#x21F2;" /*glyphicon-star*/, "glyphicon-arrow-right", "#")  : "")
-        })
-                 }
-        // .popover('show');
-      
-      /*
-        .on("shown.bs.popover", function(event) {
-          if (trace_gui)
-            console.log("shown.bs.popover", event, event.target);
-        });
-      */
-          // button is not found when show.bs.popover, but is when shown.bs.popover.
-          // Could select via id from <text> attr aria-describedby="popover800256".
-      /*
-          let deleteButtonS = d3.select("button.DeleteMap");
-          if (trace_gui)
-            console.log(deleteButtonS.empty(), deleteButtonS.node());
-          deleteButtonS
-            .on('click', menuActions.axisDelete);
-      */
-
-      /*
-          let flipButtonS = d3.select("button.FlipAxis");
-          flipButtonS
-            .on('click', menuActions.axisFlip); */
-
-      /*
-          let perpendicularButtonS = d3.select("button.PerpendicularAxis");
-          perpendicularButtonS
-            .on('click', menuActions.axisPerpendicular); */
-
-      /*
-          let extendButtonS = d3.select("button.ExtendMap");
-          if (trace_gui)
-            console.log(extendButtonS.empty(), extendButtonS.node());
-          extendButtonS
-            .on('click', menuActions.axisExtend); */
-
-    }
-
-    /*------------------------------------------------------------------------*/
-
-    /** Setup hover menus over axis child data block sub-titles.
-     * Based on similar @see configureAxisTitleMenu()
-     * @param block (Block) is the __data__ of the <tspan>-s
-     */
-    function  configureAxisSubTitleMenu(block) {
-      if (trace_gui)
-      console.log("configureAxisSubTitleMenu", block.axisName, this, this.outerHTML);
-        let node_ = this;
-      let blockR = block.block,
-      title = blockR
-        ? blockR.get('namespace') + ' ' + blockR.get('scope')
-        : block.longName();
-      const menuActions = AxisMenuActions(oa);
-      if (true /* use axis-menu.hbs, not $.popover*/) {
-        menuActions_block();
-      } else
-      if ($(node_) .popover)
-        $(node_)
-        .popover({
-          /* would like to use .axis-menu as a selector in css,
-           * but 'class' is not effective; maybe in a later version. refn :
-           * https://github.com/twbs/bootstrap/pull/23874 */
-          class : 'axis-menu',
-            trigger : "hover",
-          sticky: true,
-          delay: {show: 200, hide: 1500},
-          container: 'div#holder',
-          placement : "auto bottom",
-          title : title,
-          html: true,
-
-          content : ""
-            + iconButton("DeleteMap", "Delete_" + block.axisName, "&#x2573;" /*glyphicon-sound-7-1*/, "glyphicon-remove-sign", "#")
-            + iconButton("VisibleAxis", "Visible_" + block.axisName, "&#x1F441;" /*Unicode Character 'EYE'*/, "glyphicon-eye-close", "#", true)
-          // glyphicon-eye-open	
-        })
-        // .popover('show');
-      
-      /*
-        .on("shown.bs.popover", function(event) {
-          if (trace_gui)
-            console.log("shown.bs.popover", event, event.target);
-
-          let deleteButtonS = d3.select("button.DeleteMap");
-          if (trace_gui)
-            console.log(deleteButtonS.empty(), deleteButtonS.node());
-          deleteButtonS
-            .on('click', menuActions.blockUnview */
-      /*buttonElt , i, g*/ /*);
-
-          let visibleButtonS = d3.select("button.VisibleAxis");
-          if (trace_gui)
-            console.log(visibleButtonS.empty(), visibleButtonS.node());
-
-          visibleButtonS
-            .on('click', menuActions.blockVisible */
-     /*buttonElt , i, g*/ /*);
-
-        });*/
-    }
-    
 
     /*------------------------------------------------------------------------*/
